pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adversarial-workflow"

version = "0.6.4"

description = "Multi-stage AI evaluation system for task plans, code review, and test validation"
readme = "README.md"
authors = [
    {name = "Fredrik Matheson"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
]
keywords = ["code-review", "ai", "llm", "quality", "testing", "aider"]
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "python-dotenv>=0.19.0",
    "aider-chat>=0.86.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=3.0",
    "black>=22.0",
    "isort>=5.0",
    "flake8>=4.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.urls]
Homepage = "https://github.com/movito/adversarial-workflow"
Documentation = "https://github.com/movito/adversarial-workflow/blob/main/README.md"
Repository = "https://github.com/movito/adversarial-workflow"
Issues = "https://github.com/movito/adversarial-workflow/issues"

[project.scripts]
adversarial = "adversarial_workflow.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["adversarial_workflow*"]

[tool.setuptools.package-data]
adversarial_workflow = ["templates/*", "templates/.*", "templates/agent-context/*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
